Fixes #12912: Use giter8 for plugin template
