--- blank-template/Makefile
@@ -1,13 +1,0 @@
-#
-# make all : will build node-external-reports.rpkg.
-# make licensed : will build a license limited version of the plugin
-#
-
-FILES = # add the file that need to be package here like $(NAME)/BLANK-schema.sql
-SCRIPTS = postinst
-
-include ../makefiles/common-scala-plugin.mk
-
-target/$(NAME)/BLANK-schema.sql:
-	cp ./src/main/resources/BLANK-schema.sql target/$(NAME)/
-
