--- blank-template/build.conf
@@ -1,30 +1,0 @@
-#
-# This file defines the release information about the plugin like
-# its version and its ABI compability. 
-#
-# So version are not managed in pom.xml (safe for parent-pom version,
-# which can't be a parameter, and must be equals to rudder-branch here)
-#
-
-# Unique identifier of the plugin
-plugin-name=blank-template
-# the full name is derived from rudder-plugin-name
-plugin-fullname=rudder-plugin-${plugin-name}
-
-# Human readable short/title descrption (used for one line text)
-plugin-title-description="""TODO - <br/>see how we can use
-triple quote for multi-lines?"""
-
-# WEB, HTML description. 
-plugin-web-description=<p>TODO</p>
-
-# Plugin version. It is build as follow: A.B-x.y(.z) with:
-# - A.B: Rudder major.minor
-# - x.y(.z): plugin major.minor.micro. Micro should be omitted. When omitted, z is assumed to be 0. 
-# For the build, we split the information between two properties, rudder branch and plugin version, 
-# which must be concaneted with "-" to build the plugin version. 
-plugin-branch=1.0
-
-# rudder branch comes from parent
-plugin-version=${rudder-branch}-${plugin-branch}
-
