--- blank-template/packaging/metadata
@@ -1,11 +1,0 @@
-{
-  "type": "plugin",
-  "name": "${plugin-id}",
-  "version": "${plugin-version}",
-  "build-date": "${maven.build.timestamp}",
-  "build-commit": "${commit-id}",
-  "jar-files": [ "/opt/rudder/share/plugins/${plugin-name}/${plugin-name}.jar" ],
-  "content": {
-    "files.txz": "/opt/rudder/share/plugins"
-  }
-}
