--- blank-template/packaging/postinst
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# add post-installation script, like:
-
-# psql -U rudder -h localhost -d rudder -f /opt/rudder/share/plugins/BLANK/BLANK-schema.sql
-
