--- blank-template/pom-template.xml
@@ -1,142 +1,0 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.normation.plugins</groupId>
-  <artifactId>blank-template</artifactId>
-  <packaging>jar</packaging>
-  <!--
-    Do not change version here, update content into: build.conf
-  -->
-  <version>${plugin-version}</version>
-
-  <parent>
-    <groupId>com.normation.plugins</groupId>
-    <artifactId>plugins-parent</artifactId>
-    <version>${rudder-branch}-${parent-version}</version>
-  </parent>
-
-  <description>
-    TODO
-  </description>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>properties-maven-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <artifactId>maven-resources-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <artifactId>maven-shade-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <!-- Add plugin specific dependencies -->
-  </dependencies>
-
-  <!-- Below is an horrible if/then/else in maven. You shouldn't have anything to change here -->
-
-  <!-- Internal profile: FOR INTERNAL USE ONLY - active if -Dlimited is *not* specified. -->
-  <profiles>
-  <profile>
-    <id>internal-default</id>
-    <activation>
-      <!-- Activation via *absence* of a system property to ensure mutual exclusivity
-           of this profile with internal-limited -->
-      <property><name>!limited</name></property>
-    </activation>
-    <build><plugins>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>build-helper-maven-plugin</artifactId>
-        <version>3.0.0</version>
-        <executions>
-          <execution>
-            <id>add-source</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>add-source</goal>
-            </goals>
-            <configuration>
-              <sources>
-                <source>${basedir}/src/main/scala-templates/default</source>
-              </sources>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins></build>
-  </profile>
-  <!-- Internal profile: FOR INTERNAL USE ONLY - active if -Dlimited is *not* specified. -->
-  <profile>
-    <id>internal-limited</id>
-    <activation>
-      <!-- Activation via *presence* of a system property to ensure mutual exclusivity
-           of this profile with internal-default -->
-      <property><name>limited</name></property>
-    </activation>
-    <build><plugins>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>templating-maven-plugin</artifactId>
-        <version>1.0.0</version>
-        <executions>
-          <execution>
-            <id>filter-src</id>
-            <goals>
-              <goal>filter-sources</goal>
-            </goals>
-                <configuration>
-                  <sourceDirectory>${basedir}/src/main/scala-templates/limited</sourceDirectory>
-                  <outputDirectory>${project.build.directory}/generated</outputDirectory>
-                </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>build-helper-maven-plugin</artifactId>
-        <version>3.0.0</version>
-        <executions>
-          <execution>
-            <id>add-license-checker-source</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>add-source</goal>
-            </goals>
-            <configuration>
-              <sources>
-                <source>${project.build.directory}/generated</source>
-              </sources>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins></build>
-    <dependencies>
-      <dependency>
-        <groupId>com.normation.plugins</groupId>
-        <artifactId>plugins-common-private</artifactId>
-      </dependency>
-    </dependencies>
-    <repositories>
-    <repository>
-      <id>rudder-release-private</id>
-      <url>http://nexus.normation.com/nexus/content/repositories/private-releases/</url>
-      <layout>default</layout>
-      <snapshots><enabled>false</enabled></snapshots>
-    </repository>
-    <repository>
-      <id>rudder-snapshot-private</id>
-      <url>http://nexus.normation.com/nexus/content/repositories/private-snapshots/</url>
-      <layout>default</layout>
-      <snapshots><enabled>true</enabled></snapshots>
-    </repository>
-    </repositories>
-  </profile>
-  </profiles>
-</project>
-
