--- blank-template/src/main/resources/template/BLANKManagement.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf8" />
-  <!-- everything here will be ignored and replaced by common-layout -->
-</head>
-<body data-lift-content-id="plugin-main">
-  <div id="plugin-main" data-lift="surround?with=common-layout;at=content">
-    <head_merge>
-      <title>TODO</title>
-      <link rel="stylesheet" type="text/css" href="src/main/reources/toserve/BLANK/BLANK.css" media="screen" data-lift="ServeResource?path=BLANK/BLANK.css">
-      <script src="src/main/resources/toserve/BLANK/BLANK.js" data-lift="ServeResource?path=BLANK/BLANK.js"></script>
-    </head_merge>
-   <!-- TODO: your actual plugin content -->
-  </div>
-</body>
-</html>
-
