+++ src/main/g8/LICENSE_EXCEPTION
