+++ src/main/g8/Makefile
@@ -1,0 +1,13 @@
+#
+# make all : will build node-external-reports.rpkg.
+# make licensed : will build a license limited version of the plugin
+#
+
+FILES = # add the file that need to be package here like \$(NAME)/$plugin_name$-schema.sql
+SCRIPTS = postinst
+
+include ../makefiles/common-scala-plugin.mk
+
+target/\$(NAME)/$plugin_name$-schema.sql:
+	cp ./src/main/resources/$plugin_name$-schema.sql target/\$(NAME)/
+
