+++ src/main/g8/README.asciidoc
@@ -1,0 +1,14 @@
+Rudder plugin: $name$
+-----------------------------
+
+This project is part of Rudder - Continuous configuration for effective compliance
+
+Rudder is an easy to use, web-driven, role-based solution for IT Infrastructure
+Automation & Compliance.
+
+See: http://rudder-project.org for more information.
+
+=== Description
+
+$title_description$
+
