+++ src/main/g8/default.properties
@@ -1,0 +1,16 @@
+# the name of the plugin will be used to derive:
+# - the plugin id by using lower cases and "-": my-plugin 
+# - the package name by concatening everything: myplugin
+# - main clases with camel case: MyPluginDef, MyPluginConf...
+name = My Plugin
+version = 1.0 
+
+title_description = One line description of plugin
+web_description = <p>HTML description of plugin</p>
+
+# you should not change that
+you_should_not_change_following_variables = just hit enter
+plugin_name = $name;format="lower,hyphen"$
+plugin_pkg = $name;format="word,lower"$
+plugin_class = $name;format="Camel"$
+
