+++ src/main/g8/packaging/postinst
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+# add post-installation script, like:
+
+# psql -U rudder -h localhost -d rudder -f /opt/rudder/share/plugins/$plugin_pkg$/$plugin_class$-schema.sql
+
