+++ src/main/g8/src/main/resources/template/$plugin_class$Management.html
@@ -1,0 +1,18 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf8" />
+  <!-- everything here will be ignored and replaced by common-layout -->
+</head>
+<body data-lift-content-id="plugin-main">
+  <div id="plugin-main" data-lift="surround?with=common-layout;at=content">
+    <head_merge>
+      <title>TODO</title>
+      <link rel="stylesheet" type="text/css" href="/toserve/$plugin_pkg$/$plugin_name$.css" media="screen" data-lift="with-cached-resource">
+      <script src="/toserve/$plugin_pkg$/$plugin_name$.js" data-lift="with-cached-resource"></script>
+    </head_merge>
+   <!-- TODO: your actual plugin content -->
+  </div>
+</body>
+</html>
+
