+++ src/main/g8/src/main/resources/toserve/$plugin_pkg$/$plugin_name$.css
